move 'composeWith' tests to separate file
