+++ types/ramda/tsconfig.json
@@ -46,6 +46,7 @@
         "test/complement-tests.ts",
         "test/compose-tests.ts",
         "test/composeK-tests.ts",
-        "test/composeP-tests.ts"
+        "test/composeP-tests.ts",
+        "test/composeWith-tests.ts"
     ]
 }
